Autenticación para que no cualquier usuario pueda hacer CRUD de medicamentos agregado

+++ backend/src/controllers/medicationsController.js
@@ -16,20 +16,20 @@ const getMedications = async (req, res) => {
 
 const createMedication = async (req, res) => {
 
-const { name, dosage, description, frequency } = req.body;
+    const { name, dosage, description, frequency } = req.body;
 
-const { data, error } = await supabase
-    .from("medications")
-    .insert([
-        {
-            user_id: req.user.id,
-            name,
-            dosage,
-            description,
-            frequency
-        }
-    ])
-    .select();
+    const { data, error } = await supabase
+        .from("medications")
+        .insert([
+            {
+                user_id: req.user.id,
+                name,
+                dosage,
+                description,
+                frequency
+            }
+        ])
+        .select();
 
     if (error) {
         return res.status(500).json(error);
@@ -51,7 +51,6 @@ const updateMedication = async (req, res) => {
             description,
             frequency
         })
-        // Solo actualiza el medicamento si el id coincide y si el user_id coincide con el usuario autenticado
         .eq("id", id)
         .eq("user_id", req.user.id)
         .select();
@@ -70,7 +69,8 @@ const deleteMedication = async (req, res) => {
     const { error } = await supabase
         .from("medications")
         .delete()
-        .eq("id", id);
+        .eq("id", id)
+        .eq("user_id", req.user.id);
 
     if (error) {
         return res.status(500).json(error);
